Refactored classes out of org.ohdsi.webapi.model. Refactored paging classes into org.ohdsi.webapi.paging.
