+++ src/main/java/org/ohdsi/webapi/cohortcomparison/ComparativeCohortAnalysis.java
@@ -25,7 +25,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
-import org.ohdsi.webapi.model.CommonEntity;
+import org.ohdsi.webapi.util.CommonEntity;
 
 /**
  * @author Frank DeFalco <fdefalco@ohdsi.org>
