+++ src/main/java/org/ohdsi/webapi/cohortresults/Analysis.java
@@ -1,4 +1,4 @@
-package org.ohdsi.webapi.model.results;
+package org.ohdsi.webapi.cohortresults;
 
 
 /**
