+++ src/main/java/org/ohdsi/webapi/conceptset/ConceptSet.java
@@ -23,7 +23,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
-import org.ohdsi.webapi.model.CommonEntity;
+import org.ohdsi.webapi.util.CommonEntity;
 
 /**
  *
