--- src/main/java/org/ohdsi/webapi/model/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package corresponds to model objects in the CDM schema
- */
-package org.ohdsi.webapi.model;
