--- src/main/java/org/ohdsi/webapi/model/results/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package corresponds to model objects in the Results (also named OHDSI) schema
- */
-package org.ohdsi.webapi.model.results;
