+++ src/main/java/org/ohdsi/webapi/paging/FacetValue.java
@@ -1,4 +1,4 @@
-package org.ohdsi.webapi.model;
+package org.ohdsi.webapi.paging;
 
 public class FacetValue {
   private String value;
