+++ src/main/java/org/ohdsi/webapi/paging/FacetsResponse.java
@@ -1,4 +1,4 @@
-package org.ohdsi.webapi.model;
+package org.ohdsi.webapi.paging;
 
 import java.util.List;
 import java.util.Map;
