+++ src/main/java/org/ohdsi/webapi/service/CohortResultsService.java
@@ -15,15 +15,13 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.math.NumberUtils;
 import org.ohdsi.circe.helper.ResourceHelper;
 import org.ohdsi.webapi.cohortanalysis.CohortAnalysis;
-import org.ohdsi.webapi.cohortanalysis.CohortAnalysisGenerationInfo;
 import org.ohdsi.webapi.cohortanalysis.CohortAnalysisTask;
 import org.ohdsi.webapi.cohortanalysis.CohortSummary;
 import org.ohdsi.webapi.cohortdefinition.CohortDefinitionRepository;
 import org.ohdsi.webapi.cohortresults.*;
 import org.ohdsi.webapi.cohortresults.mapper.AnalysisResultsMapper;
-import org.ohdsi.webapi.model.CohortDefinition;
-import org.ohdsi.webapi.model.results.Analysis;
-import org.ohdsi.webapi.model.results.AnalysisResults;
+import org.ohdsi.webapi.cohortresults.Analysis;
+import org.ohdsi.webapi.cohortresults.AnalysisResults;
 import org.ohdsi.webapi.source.Source;
 import org.ohdsi.webapi.source.SourceDaimon;
 import org.ohdsi.webapi.util.PreparedStatementRenderer;
